+++ samples/web/Liuliu.Demo.EntityConfiguration/Authorization/ModuleConfiguration.cs
@@ -19,7 +19,7 @@ namespace Liuliu.Demo.EntityConfiguration.Authorization
     /// <summary>
     /// 模块信息映射配置类
     /// </summary>
-    public partial class ModuleConfiguration : EntityTypeConfigurationBase<Module, int>
+    public partial class ModuleConfiguration : EntityTypeConfigurationBase<Module, long>
     {
         /// <summary>
         /// 重写以实现实体类型各个属性的数据库配置
