+++ samples/web/Liuliu.Demo.EntityConfiguration/Identity/RoleConfiguration.cs
@@ -17,7 +17,7 @@ using OSharp.Entity;
 
 namespace Liuliu.Demo.EntityConfiguration.Identity
 {
-    public partial class RoleConfiguration : EntityTypeConfigurationBase<Role, int>
+    public partial class RoleConfiguration : EntityTypeConfigurationBase<Role, long>
     {
         /// <summary>
         /// 重写以实现实体类型各个属性的数据库配置
