+++ samples/web/Liuliu.Demo.EntityConfiguration/Identity/UserClaimConfiguration.cs
@@ -16,7 +16,7 @@ using OSharp.Entity;
 
 namespace Liuliu.Demo.EntityConfiguration.Identity
 {
-    public partial class UserClaimConfiguration : EntityTypeConfigurationBase<UserClaim, int>
+    public partial class UserClaimConfiguration : EntityTypeConfigurationBase<UserClaim, long>
     {
         /// <summary>
         /// 重写以实现实体类型各个属性的数据库配置
