+++ samples/web/Liuliu.Demo.EntityConfiguration/Identity/UserConfiguration.cs
@@ -17,7 +17,7 @@ using OSharp.Entity;
 
 namespace Liuliu.Demo.EntityConfiguration.Identity
 {
-    public partial class UserConfiguration : EntityTypeConfigurationBase<User, int>
+    public partial class UserConfiguration : EntityTypeConfigurationBase<User, long>
     {
         /// <summary>
         /// 重写以实现实体类型各个属性的数据库配置
