+++ samples/web/Liuliu.Demo.Web/Areas/Admin/Controllers/Identity/RoleController.cs
@@ -107,11 +107,11 @@ namespace Liuliu.Demo.Web.Areas.Admin.Controllers
         /// <returns>角色[用户]树数据</returns>
         [HttpGet]
         [Description("读取角色[用户]树数据")]
-        public List<UserRoleNode> ReadUserRoles(int userId)
+        public List<UserRoleNode> ReadUserRoles(long userId)
         {
             Check.GreaterThan(userId, nameof(userId), 0);
 
-            int[] checkRoleIds = _identityContract.UserRoles.Where(m => m.UserId == userId).Select(m => m.RoleId).Distinct().ToArray();
+            long[] checkRoleIds = _identityContract.UserRoles.Where(m => m.UserId == userId).Select(m => m.RoleId).Distinct().ToArray();
             List<UserRoleNode> nodes = _identityContract.Roles.Where(m => !m.IsLocked)
                 .OrderByDescending(m => m.IsAdmin).ThenBy(m => m.Id).ToOutput<Role, UserRoleNode>().ToList();
             nodes.ForEach(m => m.IsChecked = checkRoleIds.Contains(m.Id));
@@ -183,11 +183,11 @@ namespace Liuliu.Demo.Web.Areas.Admin.Controllers
         [DependOnFunction("Read")]
         [UnitOfWork]
         [Description("删除")]
-        public async Task<AjaxResult> Delete(int[] ids)
+        public async Task<AjaxResult> Delete(long[] ids)
         {
             Check.NotNull(ids, nameof(ids));
             List<string> names = new List<string>();
-            foreach (int id in ids)
+            foreach (long id in ids)
             {
                 Role role = await _roleManager.FindByIdAsync(id.ToString());
                 IdentityResult result = await _roleManager.DeleteAsync(role);
